requirements.txt
pyqt6
google-cloud-vision
google-generativeai
python-dotenv
